+++ src/OpenTelemetry.Instrumentation.AspNet/AspNetInstrumentationOptions.cs
@@ -35,8 +35,11 @@ namespace OpenTelemetry.Instrumentation.AspNet
         });
 
         /// <summary>
-        /// Gets or sets a hook to exclude calls based on domain or other per-request criterion.
+        /// Gets or sets a Filter function to filter instrumentation for requests on a per request basis.
+        /// The Filter gets the HttpContext, and should return a boolean.
+        /// If Filter returns true, the request is collected.
+        /// If Filter returns false or throw exception, the request is filtered out.
         /// </summary>
-        internal Predicate<HttpContext> RequestFilter { get; set; }
+        public Func<HttpContext, bool> Filter { get; set; }
     }
 }