+++ src/OpenTelemetry.Instrumentation.Http/HttpClientInstrumentationOptions.cs
@@ -18,6 +18,7 @@ using System;
 using System.Net.Http;
 using System.Runtime.CompilerServices;
 using OpenTelemetry.Context.Propagation;
+using OpenTelemetry.Instrumentation.Http.Implementation;
 using OpenTelemetry.Trace;
 
 namespace OpenTelemetry.Instrumentation.Http
@@ -42,17 +43,28 @@ namespace OpenTelemetry.Instrumentation.Http
         });
 
         /// <summary>
-        /// Gets or sets an optional callback method for filtering <see cref="HttpRequestMessage"/> requests that are sent through the instrumentation.
+        /// Gets or sets a Filter function to filter instrumentation for requests on a per request basis.
+        /// The Filter gets the HttpRequestMessage, and should return a boolean.
+        /// If Filter returns true, the request is collected.
+        /// If Filter returns false or throw exception, the request is filtered out.
         /// </summary>
-        public Func<HttpRequestMessage, bool> FilterFunc { get; set; }
+        public Func<HttpRequestMessage, bool> Filter { get; set; }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal bool EventFilter(string activityName, object arg1)
         {
-            return
-                this.FilterFunc == null ||
-                !TryParseHttpRequestMessage(activityName, arg1, out HttpRequestMessage requestMessage) ||
-                this.FilterFunc(requestMessage);
+            try
+            {
+                return
+                    this.Filter == null ||
+                    !TryParseHttpRequestMessage(activityName, arg1, out HttpRequestMessage requestMessage) ||
+                    this.Filter(requestMessage);
+            }
+            catch (Exception ex)
+            {
+                HttpInstrumentationEventSource.Log.RequestFilterException(ex);
+                return false;
+            }
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]