+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added Filter public API on AspNetCoreInstrumentationOptions to allow
+  filtering of instrumentation based on HttpContext.
+
 * Asp.Net Core Instrumentation automatically populates HttpRequest,
   HttpResponse in Activity custom property